ai-service/requirements.txt
flask
joblib
scikit-learn==1.6.1
psycopg2-binary
pandas
